Add files via upload
Begining of Python language I have developed my second project that is a "number loop" . use of ( for loop,if,else,elif ). It may not be a big thing for a trained software developer but for a beginner who is doing everything from scratch is a big achievement presently. This project allowed me to dive deeper into Python's capabilities, improve my coding skills.